+++ src/onegov.ballot/onegov/ballot/collections/__init__.py
@@ -1,12 +1,12 @@
 from onegov.ballot.collections.ballots import BallotCollection
 from onegov.ballot.collections.elections import ElectionCollection
-from onegov.ballot.collections.election_composites import \
-    ElectionCompositeCollection
+from onegov.ballot.collections.election_compounds import \
+    ElectionCompoundCollection
 from onegov.ballot.collections.votes import VoteCollection
 
 __all__ = [
     'BallotCollection',
     'ElectionCollection',
-    'ElectionCompositeCollection',
+    'ElectionCompoundCollection',
     'VoteCollection'
 ]