+++ src/onegov.ballot/onegov/ballot/collections/election_compounds.py
@@ -1,4 +1,4 @@
-from onegov.ballot.models import ElectionComposite
+from onegov.ballot.models import ElectionCompound
 from onegov.core.collection import Pagination
 from sqlalchemy import cast
 from sqlalchemy import desc
@@ -7,7 +7,7 @@ from sqlalchemy import extract
 from sqlalchemy import Integer
 
 
-class ElectionCompositeCollectionPagination(Pagination):
+class ElectionCompoundCollectionPagination(Pagination):
 
     def __init__(self, session, page=0, year=None):
         self.session = session
@@ -20,13 +20,13 @@ class ElectionCompositeCollectionPagination(Pagination):
     def subset(self):
         query = self.query()
         query = query.order_by(
-            desc(ElectionComposite.date),
-            ElectionComposite.shortcode,
-            ElectionComposite.title
+            desc(ElectionCompound.date),
+            ElectionCompound.shortcode,
+            ElectionCompound.title
         )
         if self.year:
             query = query.filter(
-                extract('year', ElectionComposite.date) == self.year
+                extract('year', ElectionCompound.date) == self.year
             )
 
         return query
@@ -42,16 +42,16 @@ class ElectionCompositeCollectionPagination(Pagination):
         return self.__class__(self.session, 0, year)
 
 
-class ElectionCompositeCollection(ElectionCompositeCollectionPagination):
+class ElectionCompoundCollection(ElectionCompoundCollectionPagination):
 
     def query(self):
-        return self.session.query(ElectionComposite)
+        return self.session.query(ElectionCompound)
 
     def get_latest(self):
-        """ Returns the election composites with the latest date. """
+        """ Returns the election compounds with the latest date. """
 
-        latest_date = self.query().with_entities(ElectionComposite.date)
-        latest_date = latest_date.order_by(desc(ElectionComposite.date))
+        latest_date = self.query().with_entities(ElectionCompound.date)
+        latest_date = latest_date.order_by(desc(ElectionCompound.date))
         latest_date = latest_date.limit(1).first()
 
         if not latest_date:
@@ -60,11 +60,11 @@ class ElectionCompositeCollection(ElectionCompositeCollectionPagination):
             return self.by_date(latest_date)
 
     def get_years(self):
-        """ Returns a list of years for which there are election composites.
+        """ Returns a list of years for which there are election compounds.
 
         """
 
-        year = cast(extract('year', ElectionComposite.date), Integer)
+        year = cast(extract('year', ElectionCompound.date), Integer)
         query = self.session.query
         query = query(distinct(year))
         query = query.order_by(desc(year))
@@ -72,36 +72,36 @@ class ElectionCompositeCollection(ElectionCompositeCollectionPagination):
         return list(r[0] for r in query.all())
 
     def by_date(self, date):
-        """ Returns the election composites on the given date. """
+        """ Returns the election compounds on the given date. """
 
         query = self.query()
-        query = query.filter(ElectionComposite.date == date)
+        query = query.filter(ElectionCompound.date == date)
         query = query.order_by(
-            ElectionComposite.shortcode,
-            ElectionComposite.title
+            ElectionCompound.shortcode,
+            ElectionCompound.title
         )
 
         return query.all()
 
     def by_year(self, year=None):
-        """ Returns the election composites for the current/given year. """
+        """ Returns the election compounds for the current/given year. """
 
         year = year or self.year
 
         query = self.query()
-        query = query.filter(extract('year', ElectionComposite.date) == year)
+        query = query.filter(extract('year', ElectionCompound.date) == year)
         query = query.order_by(
-            ElectionComposite.date,
-            ElectionComposite.shortcode,
-            ElectionComposite.title
+            ElectionCompound.date,
+            ElectionCompound.shortcode,
+            ElectionCompound.title
         )
 
         return query.all()
 
     def by_id(self, id):
-        """ Returns the election composite by id. """
+        """ Returns the election compound by id. """
 
         query = self.query()
-        query = query.filter(ElectionComposite.id == id)
+        query = query.filter(ElectionCompound.id == id)
 
         return query.first()