+++ src/onegov.ballot/onegov/ballot/models/__init__.py
@@ -10,7 +10,7 @@ See:
 from onegov.ballot.models.election import Candidate
 from onegov.ballot.models.election import CandidateResult
 from onegov.ballot.models.election import Election
-from onegov.ballot.models.election import ElectionComposite
+from onegov.ballot.models.election import ElectionCompound
 from onegov.ballot.models.election import ElectionResult
 from onegov.ballot.models.election import List
 from onegov.ballot.models.election import ListConnection
@@ -30,7 +30,7 @@ __all__ = [
     'CandidateResult',
     'ComplexVote',
     'Election',
-    'ElectionComposite',
+    'ElectionCompound',
     'ElectionResult',
     'List',
     'ListConnection',